// frontend/src/domains/shared/phone/generated/phone-format.ts
export interface PhoneFormat {
  groups: number[];
  maxDigits: number;
}